+++ src/main/java/insane96mcp/enhancedai/modules/base/BaseModule.java
@@ -14,6 +14,7 @@ public class BaseModule extends Module {
 	public AvoidExplosions avoidExplosions;
 	public AntiCheese antiCheese;
 	public Movement movement;
+	//public Shielding shielding;
 
 	public BaseModule() {
 		super(Config.builder);
@@ -23,6 +24,7 @@ public class BaseModule extends Module {
 		avoidExplosions = new AvoidExplosions(this);
 		antiCheese = new AntiCheese(this);
 		movement = new Movement(this);
+		//shielding = new Shielding(this);
 		Config.builder.pop();
 	}
 
@@ -34,5 +36,6 @@ public class BaseModule extends Module {
 		avoidExplosions.loadConfig();
 		antiCheese.loadConfig();
 		movement.loadConfig();
+		//shielding.loadConfig();
 	}
 }